requirements.txt
pandas
sqlalchemy
psycopg2-binary
boto3
python-dotenv
requests